Manage the RestartCause property from power-control
There were race-conditions when setting the RestartCause from
outside of power-control. This change has power-control manage
the RestartCause property directly so it can handle the race-
conditions.

The race-conditions are resolved by adding a RestartCauseRequest
property that can be set by external modules to request a restart
cause value. All requested restart causes are maintained in a set.

When the host powers-off, the set of restart causes is checked and
the highest priority cause is set as the RestartCause.

When the host powers-on, the set of restart causes is cleared for
the next set of causes to be recorded.

Tested:
Triggered a watchdog and confirmed that the request was added to the
set and that based on priority, watchdog was set as the RestartCause
when the system restarted.

Triggered a software, command, button, and power restore restart
types and confirmed that the restart cause was correctly reported.

Change-Id: If968d434ec3d827e82e033bd528acaa88f625d1e
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>