Remove shut down on VR watchdog timeout
The BMC should not force the system off based on a power good
timeout.

This change removes the shutdown and the "failed transition
to On" state which only existed to handle the shutdown.
Now on errors during power-on, BMC will log the event and
go to the Off state.  The Off state will now also handle
getting a power good event and going straight to On.

Tested:
Forced a power good timeout and confirmed that the BMC and
system both end up in the On state.

Change-Id: If4574a01ae821017604ca8ac0ec54b4d281b4d3b
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>