Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > > Patch Set 2:
>  > >
>  > > (1 comment)
>  > 
>  > On the mailing list discussion with Ed Tanous, he suggested to use
>  > compile time flag rather than config file setup. @Jason kindly
>  > share your comments on this.
> 
> I would simply keep usage simple based on configuration rather than adding flag as well as config.

I think his comment may be based on that this change feels very platform-specific so it should be behind a compile flag to remove it for platforms that don't use it.  Is a GPIO to control slot AC power a standard implementation?

I think I'm okay without the compile flag, but if different platforms have different methods to execute this functionality, we may need to come up with a more generic approach to add or modify behavior.  The issue with that is that is what phosphor-state-manager is for.

Patch-set: 2
Label: Code-Review=0