Update patch set 3
Patch Set 3:

Can you please have a discussion on the mailing list with Vijay about how these two features, and the design he proposed are intended to work together.  I suspect you two are working on the same patch;  If you're not, I suspect documentation is required to determine when to to use which one.

Patch-set: 3
Label: Code-Review=0