Use phosphor::logging instead of std::cerr
After moving part of the logs to phosphor::logging, it caused
the messages to get logged out of order depending on if they
came from phosphor::logging or std::cerr.

This change moves all logs to phosphor::logging so they all
appear in the correct order again.

Also fixed a minor typo.

Tested:
Performed a power cycle and confirmed that the logs all look
as expected.

Change-Id: I25f677097d2f6552e09536c3078bb08e1d985986
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>