Update patch set 1
Patch Set 1: Code-Review-1

This is kind of a big change, so we should be careful here.  I think there is something bigger going on because the system should quickly move from transitionToOff to Off.

The expected signal change to trigger that transition is Power Supply Power OK de-assert.  If the power supplies are on, then the chassis state is still on.

Patch-set: 1
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=-1