Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi!
> As I understand this commit purpose was mainly to get the x86 system working with the openbmc-test-automation framework.
> I've recently commited several changes to the openbmc-test-automation repo and they all are now merged upstream.
> 
> Example from the updated README.md:
> ```
> By default openbmc-test-automation framework assumes that host CPU is based on the POWER architecture. If your host CPU is x86 add `-v PLATFORM_ARCH_TYPE:x86` variable setting to your CLI commands.
> ```
> With this option enabled openbmc-test-automation framework simply ignores BootProgress interface in its state checks and all tests seem to be working correctly.
> 
> Please check it out when you'll have a chance.

I was using `tox` to run the HW CI, that the `-v PLATFORM_ARCH_TYPE:x86` is not a valid option.
Instead, I tried to run with environment defined as below:

 OPENBMC_HOST=${OPENBMC_HOST} OPENBMC_PASSWORD=${OPENBMC_PASSWORD} SSH_PORT=22 PLATFORM_ARCH_TYPE=x86 tox -e default -- --argumentfile test_lists/HW_CI tests redfish

But it does not work, that it still checks the boot progress and operation system properties.

Could you kindly let me know how to let the test-automation to pick the x86 json?

Patch-set: 2