Add multiple host Slot AC Power Support
This change provides Slot AC Power feature for the system which has Slot
AC Power support. To enable this feature, the compiler
flag 'SLOT_AC_POWER' must be defined from the platform side.

The Chassis properites On, Off and PowerCycle will provide the Slot
AC Power On, Off and PowerCycle for the platform with the compiler flag
'SLOT_AC_POWER'

TESTED: Tested and verified in YosemiteV2

Signed-off-by: Naveen Moses <naveen.mosess@hcl.com>
Change-Id: I21b9326962099dfa59ff749d753121f74830db57