Update patch set 3
Patch Set 3:

(5 comments)

Overall it looks okay.  It might be interesting to see if it's possible to put the D-Bus and GPIO init stuff in separate header files, but that could be done as a separate clean-up task.

Patch-set: 3
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>