Update patch set 8
Patch Set 8:

Can you please provide more detail on the issue you see that is prompting this change?

The powerState is already initialized starting on line 3351.  It is initialized based on the state of psPowerOK.  Are you seeing that the power state is not correctly initialized based on the state of psPowerOK?

Patch-set: 8
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>