Update patch set 8
Patch Set 8:

(2 comments)

> Patch Set 8:
> 
> Can you please provide more detail on the issue you see that is prompting this change?
> 
> The powerState is already initialized starting on line 3351.  It is initialized based on the state of psPowerOK.  Are you seeing that the power state is not correctly initialized based on the state of psPowerOK?

In the case of SIO mode, the code is looking for psPowerOK and sioPowerGood. psPowerOK is not good enough by itself as initialized. 

Roughly if psPowerOK is good, the systems is declared as On but the code is waiting for sioPowerGood to be triggered.

The issue is that on ProLiant, if the host power policy is set to on, sioPowerGood is set to high before the BMC is started, and doesn't trigger any assert event leading the code to wait for ever for such event. So that is why I added that initial test and trigger the event by hand when gpios are initialized.

Patch-set: 8