Update patch set 8
Patch Set 8:

(2 comments)

> Patch Set 8:
> 
> (2 comments)
> 
> > Patch Set 8:
> > 
> > Can you please provide more detail on the issue you see that is prompting this change?
> > 
> > The powerState is already initialized starting on line 3351.  It is initialized based on the state of psPowerOK.  Are you seeing that the power state is not correctly initialized based on the state of psPowerOK?
> 
> In the case of SIO mode, the code is looking for psPowerOK and sioPowerGood. psPowerOK is not good enough by itself as initialized. 
> 
So far, this has been treated as a corner case.  We can consider adding extra logic here to handle both signals, but we would need to understand all the error cases and how to handle them.

> Roughly if psPowerOK is good, the systems is declared as On but the code is waiting for sioPowerGood to be triggered.
> 
> The issue is that on ProLiant, if the host power policy is set to on, sioPowerGood is set to high before the BMC is started, and doesn't trigger any assert event leading the code to wait for ever for such event. So that is why I added that initial test and trigger the event by hand when gpios are initialized.

Can you show where you are getting stuck waiting forever?  I added a couple comments to the code where the state is initialized to "on" when psPowerOk is asserted.  Then, since it is in the "on" state, the sioPowerGoodAssert event that you trigger is not handled and takes no action.

Is your system in a different state after initialization?

Patch-set: 8
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>