Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > ok, my initial journalctl output with not containing all the logging info. I recompiled with the proposed fix
> > 
> > root@dl360poc-9440c941b11a:~# systemctl status xyz.openbmc_project.Chassis.Control.Power.service
> > * xyz.openbmc_project.Chassis.Control.Power.service - Intel Power Control
> >      Loaded: loaded (/lib/systemd/system/xyz.openbmc_project.Chassis.Control.Power.service; enabled; vendor preset: enabled)
> >      Active: active (running) since Mon 2021-08-23 23:29:00 UTC; 3min 16s ago
> >    Main PID: 220 (power-control)
> >      CGroup: /system.slice/xyz.openbmc_project.Chassis.Control.Power.service
> >              `-220 /usr/bin/power-control
> > 
> > Aug 23 23:28:52 dl360poc power-control[220]: RESET_OUT set to 0
> > Aug 23 23:28:52 dl360poc power-control[220]:  NMI Source Property Monitor
> > Aug 23 23:28:52 dl360poc power-control[220]: Host0: Moving to "Off" state
> > Aug 23 23:28:53 dl360poc power-control[220]: powerStateOff: SIO power good assert event received
> > Aug 23 23:28:53 dl360poc power-control[220]: Host0: Moving to "On" state
> > Aug 23 23:28:53 dl360poc power-control[220]: SIO Power Good asserted during initialization as associated GPIO is HIGH
> > Aug 23 23:28:53 dl360poc power-control[220]: Initializing power state. POH timer started
> > Aug 23 23:29:00 dl360poc systemd[1]: Started Intel Power Control.
> > Aug 23 23:30:48 dl360poc-9440c941b11a power-control[220]: powerStateOn: power supply power OK assert event received
> > Aug 23 23:30:48 dl360poc-9440c941b11a power-control[220]: No action taken.
> > root@dl360poc-9440c941b11a:~# 
> > 
> > You can see the SIO power good event is triggered. Without that fix, it will not work on ProLiant
> 
> So, does that mean that your SIO power good is asserted before power supply power OK?

It could happens when we are restoring power if the mandate is to be on. Our BMC is holding the host from running up to the time we didn't had ensure that the ROM is ready to be delivered to the PCH and has been validated. We assert (in fact it is soon ok) the power supply ok later into the boot process while sio power is ok. roughly both signals where up in the right sequencing but one of them is caught by our cpld/boot logic to wait that the system ends up in the right state. the current code logic is assuming a sequencing which is not right on proliant. the proliant sequencing might be wrong but this is the way it is currently and it definitely need to enforce the triggering of the sio signal during power up otherwise, if the restore power policy is on it will fail.

Patch-set: 8