Update TimerMap names to match the config file
The TimerMap names are compared against the config file entries
to update the values; however, the names don't match between the
TimerMap and the config file, so no timing values from the config
file are used.

This updates the TimerMap to match the names in the config file
so the config file timing values can be used.

Tested:
Changed the PowerCycleMs value to 15000 and confirmed that the
system stayed off for 15 seconds on a power cycle.

Change-Id: Ia273471fca36303a33572eeedff9079a18ad3569
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>