Update patch set 11
Patch Set 11:

I have updated the code to support as an input to polarity ActiveHigh or ActiveLow. This is then translated to a boolean value inside the code, and if it is not present it reports a failure. I have also updated the default json file to follow original design format, Hope this is answering to all the open questions we might have ;)

Patch-set: 11