Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> If this is something we need to do every time we transition to the waitForSioPowerGood state, should we maybe factor it out into a helper function like
> 
> ```
> static void startSioPowerGoodWait()
> {
>   sioPowerGoodWatchdogTimerStart();
>   setPowerState(PowerState::waitForSioPowerGood);
> }
> ```
> 
> and just call that at each site, so that future changes are less likely to omit it?

I added another review where I added helpers for each of these timer and setPowerState pairs.  It introduced an issue where one timer is used for two different setPowerStates, so it will require two helpers.  It's starting to feel like a bit of overkill, but we can discuss in the other review if it will provide much benefit.

Patch-set: 2