Add helper to assert GPIOs for a time duration
This change adds a helper that uses the defined polarity to
assert a GPIO for a time duration.  This helper acts as a
middle layer to abstract the actual value from both the caller
and the setGPIOOutputForMs() function.

Tested:
Confirmed that ipmitool power off, on, cycle, and reset all
behave correctly.

Change-Id: Ie3a824b1e23474fc839ce1fc1f6af6283f7aa39d
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>