Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > > Patch Set 4: Code-Review+1
> > > > > 
> > > > > Looks good to me.  I'm giving +1 for a little bit for others to give any feedback before I promote it to +2.
> > > > 
> > > > Change looks good to me. Is it tested? I have to take up this patch and test it.
> > > 
> > > I tested it in combination with a pending config for e3c246d4i (see https://gerrit.openbmc-project.xyz/c/openbmc/openbmc/+/46642) and it worked with those GPIOs; I'm not set up with a system to test the DBus signal handling paths though, so if you could verify that it doesn't break those that'd be great.
> > 
> > @Naveen are you planning to test the D-Bus functionality with this change?
> 
> Yes I will test this change and update on it.

Hi All,
To test this feature in our platform we have to enable another platform specific feature which is work in progress. so after verifying that feature we can test and confirm this review changes. So it will take a week or two. meanwhile if this feature is implementation good we can merge it.
Thanks.

Patch-set: 4
Reviewer: Gerrit User 1001093 <1001093@adbd0a64-1f21-4d83-b585-671fe73cb6e4>