Update patch set 5
Patch Set 5:

> Patch Set 1:
> 
> This looks like a good fix -- while we're messing with this code though, might we perhaps consider dropping the suffixes and normalizing everything to float seconds instead of a mix of seconds and milliseconds?  (c.f. https://gerrit.openbmc-project.xyz/c/openbmc/dbus-sensors/+/44463/6#message-10c9d084915e6ca4774b5c7ff8de2abff584aed7)

Yes, I think this is worth considering, but I don't know if it fits as part of this change.  This change focuses on fixing a bug where the config file values are ignored.  The config file names aren't changed here, so I think it would work better as a separate commit to update the names and normalize to float seconds.

Patch-set: 5