Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I think it looks like an improvement -- makes it easier to get things right when making changes.  My one bikesheddy quibble would be with the function naming; the "request" nomenclature is currently used with `requestGPIOEvents`, which is more of an "open" operation than a "set" operation, so I think distinguishing the naming of these functions a bit more from that might be good.  Maybe just `assertGPIO()` and `deassertGPIO()`?

I struggled with naming this, too.  I had thought about just assertGPIO() and deassertGPIO(), but I felt like that was hiding the fact that the function does a request() call on the GPIO to claim ownership of it.  However, I see that it's confusing when it calls the "set" function instead of a "request".  I will change it.

Patch-set: 1