Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > (1 comment)
> > > 
> > > I think it looks like an improvement -- makes it easier to get things right when making changes.  My one bikesheddy quibble would be with the function naming; the "request" nomenclature is currently used with `requestGPIOEvents`, which is more of an "open" operation than a "set" operation, so I think distinguishing the naming of these functions a bit more from that might be good.  Maybe just `assertGPIO()` and `deassertGPIO()`?
> > 
> > I struggled with naming this, too.  I had thought about just assertGPIO() and deassertGPIO(), but I felt like that was hiding the fact that the function does a request() call on the GPIO to claim ownership of it.  However, I see that it's confusing when it calls the "set" function instead of a "request".  I will change it.
> 
> Oh, well now I feel a bit silly -- I actually hadn't realized that those were (indirectly) calling gpiod::line::request().  Dumb question time: what's the reason we re-open those every time we set them instead of opening them during initialization and holding on to them?

It depends on the GPIO.  For generic GPIOs like the NMI Output pin, we can request and hold those permanently.  For the Power and Reset Output pins, those are pass-through pins where the underlying pinctrl driver is configured to automatically disconnect the pass-through when the pins are claimed.  So, to manually control the pins, we have to claim them.  But to allow them to be controlled by the front-panel power and reset buttons, we have to release them to re-enable the pass-through.

Patch-set: 2