Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Sujoy,
> 
> The patch looks alright, though I wonder if this might be a good opportunity to go a bit further and convert all the settings to (float) seconds for consistency with other things as has come up in some recent discussion: https://gerrit.openbmc-project.xyz/c/openbmc/x86-power-control/+/46697/6#message-1804d7f4002d6d54efc8c08cb45ae3832bde4287
> 
> Also, if we're going to be making backwards-incompatible changes like this to the config file, it'd be good to also update the existing configs in the tree at the same time so they don't break.
> 
> 
> Zev

Hi Zev,
Thank you for your comment on this. The timer values in question are all in milliseconds and they describe the timer usage accurately. We are using chrono helper types and they all take signed integer as value - not float. We could experiment further but at this point we're little hesitant to make big changes as this may impact platform stability. We will review your suggestion at a later date. For now, it will be helpful and appreciated if you please review the current patch and let me know if this can be submitted,

Best regards,
Sujoy.

Patch-set: 1