Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > We are using chrono helper types and they all take signed integer as value - not float.
> 
> Sure -- that seems like an implementation detail though, it doesn't have to dictate our config design.
> 
> > We could experiment further but at this point we're little hesitant to make big changes as this may impact platform stability. We will review your suggestion at a later date.
> 
> Can I ask who "we" means here?  (Mostly I'm wondering if it includes Jason or Vernon.)
>
Hi Zev.  Sorry for the confusion.  As I understand it, our end goal is to normalize the config file timing values to simplify the config file input?

With that in mind, we looked at the change to float seconds compared to milliseconds, and it was much less intrusive to normalize to milliseconds.  So, we decided to use that as the starting point for this review discussion.

I like the idea of normalizing all of OpenBMC timing configuration to float seconds.  Do you know if that is what is happening and other OpenBMC repos are also normalizing timing configs to float seconds?

Patch-set: 1
Reviewer: Gerrit User 1000276 <1000276@adbd0a64-1f21-4d83-b585-671fe73cb6e4>