Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > We are using chrono helper types and they all take signed integer as value - not float.
> > 
> > Sure -- that seems like an implementation detail though, it doesn't have to dictate our config design.
> > 
> > > We could experiment further but at this point we're little hesitant to make big changes as this may impact platform stability. We will review your suggestion at a later date.
> > 
> > Can I ask who "we" means here?  (Mostly I'm wondering if it includes Jason or Vernon.)
> >
> Hi Zev.  Sorry for the confusion.  As I understand it, our end goal is to normalize the config file timing values to simplify the config file input?
> 
> With that in mind, we looked at the change to float seconds compared to milliseconds, and it was much less intrusive to normalize to milliseconds.  So, we decided to use that as the starting point for this review discussion.
> 
> I like the idea of normalizing all of OpenBMC timing configuration to float seconds.  Do you know if that is what is happening and other OpenBMC repos are also normalizing timing configs to float seconds?

I don't know offhand what other configurable time intervals presently exist in OpenBMC, but at least within entity-manager/dbus-sensors there's been some conscious effort to move in that direction: https://gerrit.openbmc-project.xyz/c/openbmc/dbus-sensors/+/44463/6#message-10c9d084915e6ca4774b5c7ff8de2abff584aed7

I think that'd be a good thing to aim for in x86-power-control.  It obviates the need for unit suffixes (we can use just `foo` instead of `fooMs` or whatever), and thus the question of whether it makes more sense to aim for all things to use the same unit or to have the unit be something that makes sense for a particular setting in isolation -- e.g. in this case the graceful shutdown timeout value is more naturally expressed in seconds, which makes it inconsistent with the other settings that are all in milliseconds, but if we aim for consistency (as in this patch), we sacrifice a bit on comprehensibility/ease-of-use for this specific setting.  Float seconds is a single format that's fairly intuitive to read/write while used consistently across a variety of timescales -- seeing 0.001 is pretty easily comprehensible as "a millisecond" and 3600 is likewise easily understood as "an hour", whereas it's (IMO) somewhat less obvious that 60000 means "one minute" (for something in milliseconds).

If there's agreement that that's what we'd like to move toward, the other consideration IMO is the amount of churn in the process of getting there -- I think it'd be good to minimize the number of times we make breaking changes to the config syntax along the way, so I'd be reluctant to proceed with this change as-is if we're just planning on changing it again in the near future.

(Adding Ed to the CC list since he was involved in the E-M time-units discussion too.)

Patch-set: 1
CC: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Reviewer: Gerrit User 1000733 <1000733@adbd0a64-1f21-4d83-b585-671fe73cb6e4>