Add BmcReady signal
BmcReady signal is a special signal that can be found on AMD boards.
It is a level GPIO, and if it is not activated onboard CPLD wouldn't
perform any power operations.
Therefore if this GPIO is present in the JSON configuration activate it
in the main function.

Tested on the AMD EthanolX CRB. With this change it is possible to get
rid of custom AMD patches and use just 'power-config-host0.json' file.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>
Change-Id: I3a21c2e5214dfcf05cdae8efc2ebbb21c4b4e080