Add HpmStbyEn signal
HpmStbyEn signal is a special signal that can be found on AMD boards.
It indicates that all DC-SCM STBY power rails are enabled and good.
The signal enables STBY power rails on the HPM.
It is a level GPIO, and if it is not activated, onboard CPLD wouldn't
relinquish control of power control GPIOs to BMC making it impossible
to power on the host.
Therefore if this GPIO is present in the JSON configuration activate it
in the main function.

Tested on the AMD EthanolX CRB

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>
Change-Id: I3a21c2e5214dfcf05cdae8efc2ebbb21c4b4e080