Add option to ignore soft resets before POST
In some cases it is expected that the host will
send a soft reset.  This can overwrite previous power state
changes which we may want to keep.

Tested:
Set restore policy to always on. AC cycle system.
Check restart cause is always on, and not overwritten as soft reset

Once host has posted, do warm reset through host and
check that restart cause is soft reset.

Change-Id: I0bd861f25680a11fc2e7eb77088f3009bcf8b03b
Signed-off-by: Matt Simmering <matthew.simmering@intel.com>