WIP: Port to libgpiod 2.0
This is totally untested, but ports x86-power-control to libgpiod 2.0.

https://gerrit.openbmc.org/c/openbmc/openbmc/+/61485

The changes can be summarized as:
line_request is no longer default constructible, so wrap in an optional<>

settings are ported to the new mechanism.

get/set_values now accepts an enum instead of an int.

Ideally at some point in the future we'd take advantage of being able to
merge all the line requests for a single chip into one FD and callback,
but that's a much more invasive change.

Change-Id: Iff6a72066a8e25e53763c18de38bd864933a86f1
Signed-off-by: Ed Tanous <edtanous@google.com>