config option: power button uses hw powercontrol
new configuration option 'UsesHWPowerControl'
is used to determine if the power button is connected
to the power control logic on the board.

On some Tyan S5549 boards, the power button
is not connected to the hardware logic for power control.

This can be fixed by performing the same logic which would
be triggered in case of dbus power control, when the button
is pressed.

The default behavior remains the same as before.

Tested: On a Tyan S5549. Power Button now works as expected.

Change-Id: Ic5c486cb93dd60fc21f1214a6e7ee803de960cf4
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>