Fix post complete status error when using D-Bus
In GPIO mode, the post complete status should be compared with the
polarity. But hardcode now, generally the polarity is 0.

In D-Bus mode, the post complete status without polarity, generally we
consider the status is standby when the property is true, otherwise the
status is inactive.

But now, we consider the status is inactive when the property is true.

This patch fix 2 issues:
1. remove hardcode polarity.
2. fix the status error when using D-Bus.

Change-Id: Ib50ea9e15fdd1b2754018c06bd2a791848a8c5bf
Signed-off-by: Jian Zhang <zhangjian.3032@bytedance.com>