build: use allowed over enabled
Meson feature options are typically in a tri-state of enabled, disabled, or
auto.  The enabled and disabled functions on an option (from `get_option`)
no longer return true for auto features.  Instead, the expectation is to
use `allowed()` which is true for both enabled and auto.

Switch all uses of `enabled` to `allowed`.

Change-Id: I70a7a959e056eca2c58c4656dd9aa323ef0e3984
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>