Require polarity config parameter on dbus signals
Previously the "DBUS" config type neither required nor respected a
polarity setting, meaning dbus-based signals needed to adhere to an
unspecified, signal-specific "default" polarity implicitly expected by
the existing code.

In order to more flexibly accommodate different dbus signal sources (and
make things clearer and more explicit), let's (a) record the parameter
in our ConfigData object and (b) make it mandatory as it is for direct
GPIO signals.  (And update the README to include it in the examples
while we're at it.)

Note that for compatibility this will require a small update to existing
configs that use the "DBUS" config type to state their polarities
explicitly.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: I529fd6362db9a49b859f97a231c09d23a782c23d