Update PostComplete polarity
After recent changes adding support for GPIO polarity in the config file
and using only active-high polarity for D-Bus, the PostComplete polarity
was left out of sync in a couple places.

This fixes PostComplete, so the correct polarity is used in all cases.

Tested:
Confirmed in D-Bus mode that PostComplete correctly initializes the OS
State and correctly updates the OS State on PropertiesChanged events.

Change-Id: Iae626de79b5f13292372081f14ca5ca75420a1c2
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>