implement Boot Progress interface
This is a naive and rather limited implementation to provide the dbus
interface.

It can be used if you just need a simple implementation to e.g. pass
test suites.

Tested: bmcweb recognizes the dbus interfaces and returns the expected
properties in the json response.

"BootProgress": {
"LastState": "OSBootStarted",
"LastStateTime": "2024-10-16T16:17:01.720000+00:00"
},

Change-Id: I2198005bb2b83d778d54510e30191677651dc801
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>