Change name to just power OK
A system power OK signal may not necessarily be tied to the power
supplies, so rename everything to remove reference to the power supply
and keep just "power OK".

Tested:
Confirmed that ipmitool power on, off, and cycle all work as expected
and there is no mention of power supply in the logs.

Change-Id: I62370f2666602525c6b4564ccff8d18b703744f4
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>